Flash File System: Fight the linker
Somehow, __embed_start and __embed_end are not defined properly. I don't
know why.

If I build two C files one of which defines __embed_start and the other
references __embed_start, it works.

Doing the same thing with my custom file doesn't work. Maybe I messed up
the generation somehow?

+++ Tools/FileSystem.cpp
@@ -164,7 +164,7 @@ void FileSystem::finalize()
         .st_shndx = static_cast<uint16_t>(*m_tab_index),
     });
     m_generator.symtab().add_symbol("__embed_end", Elf32_Sym {
-        .st_value = m_tab_stream.size(),
+        .st_value = static_cast<uint32_t>(m_tab_stream.size()),
         .st_size = 0,
         .st_info = ELF32_ST_INFO(STB_GLOBAL, STT_OBJECT),
         .st_other = STV_DEFAULT,
